+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Embedded Graphics is a `no_std` library for adding graphics features to display 
 ### Added
 
 - Added [hub75](https://crates.io/crates/hub75) driver to list of known supported drivers.
+- tinbmp: Added `.into_iter()` support to the `Bmp` struct to get an iterator over every pixel in the image.
 
 ### Changed